+++ crates/swc/tests/tsc-references/optionalChainingInParameterInitializer(target=es5).2.minified.js
@@ -1,5 +1,6 @@
 //// [optionalChainingInParameterInitializer.ts]
-var _a, a = function() {};
+var a = function() {};
 !function() {
+    var _a;
     arguments.length > 0 && void 0 !== arguments[0] ? arguments[0] : null === (_a = a()) || void 0 === _a || _a.d;
 }();