+++ crates/swc_ecma_transforms_compat/tests/optional-chaining/issue-1130/1/output.js
@@ -1,2 +1,2 @@
-var _data_filter_map, _this, _data_filter, _this1;
-const result = (_this1 = data) === null || _this1 === void 0 ? void 0 : (_data_filter_map = (_this = (_data_filter = _this1.filter) === null || _data_filter === void 0 ? void 0 : _data_filter.call(_this1, (item)=>Math.random() > 0.5)).map) === null || _data_filter_map === void 0 ? void 0 : _data_filter_map.call(_this, (item)=>JSON.stringify(item));
+var _data;
+const result = (_data = data) === null || _data === void 0 ? void 0 : _data.filter((item)=>Math.random() > 0.5).map((item)=>JSON.stringify(item));