+++ crates/swc_ecma_transforms_typescript/tests/fixture/issue-7539/input.tsx
@@ -1,0 +1,3 @@
+/* @jsx h */
+/* @jsxFrag null */
+<></>;